Format CLI SRP materialization helpers
