+++ crates/uselesskey-cli/src/srp/mod.rs
@@ -1,0 +1,5 @@
+//! Single-responsibility implementation modules for CLI helper APIs.
+
+pub mod export;
+pub mod manifest;
+pub mod materialize;
